refactor(trusty-agents): split context/manager trim machinery into context/trim; fix Test: pointers
Two deterministic CI gates burned by the four review rounds:

1. 500-SLOC file cap (issue #610): context/manager.rs reached 572 SLOC. Split
   the trim/truncation machinery — trim_to_budget (as an `impl ContextManager`
   block), water_fill_cap, recency_window_start, clamp_recency_to_pairings,
   truncate_message_content, estimate_tokens, and the tool_call pairing helpers
   — verbatim into a sibling module context/trim.rs (212 SLOC), with all their
   tests in context/trim/tests.rs (320 SLOC; basename tests.rs → 1500 test cap).
   manager.rs is now the thin public surface (context_window, TrimOutcome,
   ContextManager + new) at 47 SLOC. Pure move, zero behavior change.

2. Test:-doc-pointer lint (issue #2458): three pointers had drifted across the
   rounds — `trim_respects_protected_count` → the real
   `trim_respects_protected_count_greater_than_len`, and
   `truncate_message_content_shortens_string` /
   `truncate_message_content_ignores_non_string` → the real
   `truncate_message_content_string_vs_non_string`. Fixed to the actual names;
   swept every pointer in the moved files (all now resolve to a real fn/mod in
   the crate).

Gates: cargo check + clippy -p trusty-agents --all-targets -D warnings + cargo
test -p trusty-agents (2277 passed, 0 failed — same count, no tests lost) +
cargo fmt --all --check all clean; scripts/check_line_cap.sh reports 0
violations; every Test: pointer in the changed files resolves against the crate
fn/mod name set (the exact check scripts/check_test_pointers.sh performs).
